+++ src/main/java/domain/repository/DatabaseConnector.java
@@ -1,4 +1,4 @@
-package persistence;
+package domain.repository;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
