+++ src/test/java/domain/repository/JdbcGameRepositoryTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.DatabaseConnector;
 
 import java.sql.Connection;
 import java.sql.Statement;
